feat(vqb): refacto of aggregation management + added keep-original-granularity
