+++ server/weaverbird/backends/sql_translator/steps/aggregate.py
@@ -41,7 +41,7 @@ def translate_aggregate(
         query_name=query_name,
         transformed_query=f'{query.transformed_query}, {query_name} AS {query_string})',
     )
-    print("new_query.transformed_query: ", new_query.transformed_query)
+
     new_query.metadata_manager = query.metadata_manager
     new_query.selection_query = build_selection_query(
         query.metadata_manager.retrieve_query_metadata_columns(), query_name
